Switch amount from double to big decimal

+++ src/main/java/com/varunsdave/personafinance/budget/service/fileupload/FileUploadService.java
@@ -6,6 +6,7 @@ import com.varunsdave.personafinance.budget.service.transactions.TransactionServ
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
@@ -23,7 +24,7 @@ public class FileUploadService {
             Transaction t = new Transaction(accountId);
             t.setDescription(transaction.getDescription());
             t.setType(transaction.getType().toLowerCase());
-            t.setAmount(transaction.getAmount());
+            t.setAmount(BigDecimal.valueOf(transaction.getAmount()));
             t.setTransactionDate(new Date(transaction.getTransactionDate()));
             return transactionService.create(t, accountId);
         }).collect(Collectors.toList());
